cleaned Interpreter
